+++ Gruntfile.js
@@ -1,7 +1,7 @@
 /*global module:false*/
 module.exports = function(grunt) {
 
-  var version = '1.7.5';
+  var version = '1.7.7';
   // before release:
   // update bower.json, package.json
   // after release: